The Version Update
Civilian Industries - V 0.2.5
We now have a version in game. This is good, because it means I can support prior saves better.
This is bad, because I have no way to check for prior game versions.

So this'll break saves. But it'll be the last time for the 0.2 series.
For the forseeable future, expect every 0.0.X change to be save safe, 0.X.0 to be conditionally save safe, and X.0.0 change to bluescreen your computer.

Fixed a bug where Cargo Ships were not being correctly forgotten by the mod on death.

Overhauled Scatter Waves into Wormhole Raids.
-Every 15-30 minutes, the AI will launch an attack on a section of your trade network.
-You will be given a warning 60-120 seconds before the attack begins.
-This warning comes in the form of a bunch of wormholes appearing on planets around their primary target.
-Upon launching the attack, a large quantity of raid ships will attempt to rush down and kill every cargo ship on these planets.