0.4.1
Fixed a bug where every ship used to raid with was semi-permanently removed from the AI as a choice.

Fixed a potential future bug for militia ships that fail to find an optimal place to path to for an Outpost location.

Removed a bit of debugging code left in from the fleet-removing overhaul. (I missed one related to trading, whoops.)